No metric selected by default.
- Fixes metric keyboard selection order
