+++ client/app/scripts/components/metric-selector-item.js
@@ -1,25 +1,7 @@
 import React from 'react';
 import classNames from 'classnames';
 import { selectMetric, lockMetric, unlockMetric } from '../actions/app-actions';
-
-
-const METRIC_LABELS = {
-  docker_cpu_total_usage: 'Container CPU',
-  docker_memory_usage: 'Container Memory',
-  host_cpu_usage_percent: 'Host CPU',
-  host_mem_usage_bytes: 'Host Memory',
-  load1: 'Host Load 1',
-  load15: 'Host Load 15',
-  load5: 'Host Load 5',
-  open_files_count: 'Process Open files',
-  process_cpu_usage_percent: 'Process CPU',
-  process_memory_usage_bytes: 'Process Memory'
-};
-
-
-export function label(m) {
-  return METRIC_LABELS[m.id];
-}
+import { label } from '../utils/data-utils';
 
 
 export class MetricSelectorItem extends React.Component {
