+++ client/app/scripts/utils/data-utils.js
@@ -41,6 +41,24 @@ function getNextValue(keyValues, maxValue) {
   return v;
 }
 
+export const METRIC_LABELS = {
+  docker_cpu_total_usage: 'Container CPU',
+  docker_memory_usage: 'Container Memory',
+  host_cpu_usage_percent: 'Host CPU',
+  host_mem_usage_bytes: 'Host Memory',
+  load1: 'Host Load 1',
+  load15: 'Host Load 15',
+  load5: 'Host Load 5',
+  open_files_count: 'Process Open files',
+  process_cpu_usage_percent: 'Process CPU',
+  process_memory_usage_bytes: 'Process Memory'
+};
+
+
+export function label(m) {
+  return METRIC_LABELS[m.id];
+}
+
 
 const METRIC_FORMATS = {
   docker_cpu_total_usage: 'percent',
